requirements.txt
# Core dependencies
python-binance>=1.0.16
requests>=2.28.1
pandas>=1.5.0
numpy>=1.21.0

# Technical analysis
TA-Lib>=0.4.0

# Web server
Flask>=2.3.0
python-dotenv>=0.21.0

# Configuration
PyYAML>=6.0

# Utilities
python-dateutil>=2.8.0
pytz>=2022.1

